adicionando interface para recursos
